630: * resolve merge conflict * new error navigaiton to created surveys not working

+++ apps/api/src/common/file-attachment/attachment.service.ts
@@ -44,6 +44,10 @@ class AttachmentService {
 
   async deleteTemporaryDirectory(userId: string) {
     const destination = `${this.filePath}/${TEMPORARY_ATTACHMENT_DIRECTORY_NAME}/${userId}`;
+    const exists = await FilesystemService.checkIfFileExist(destination);
+    if (!exists) {
+      return;
+    }
     await this.fileSystemService.deleteDirectory(destination);
   }
